Changed float to double, update EF Core to 3.1.2

+++ Contracts/Entities/IWorkout.cs
@@ -6,12 +6,12 @@
     public interface IWorkout : IEntity
     {
         DateTime WorkoutDate { get; set; }
-        float Distance { get; set; }
-        float Duration { get; set; }
+        double Distance { get; set; }
+        double Duration { get; set; }
         //ICollection<IWorkoutInterval> Intervals { get; set; }
         string Name { get; set; }
         string Note { get; set; }
-        float Pace { get; set; }
+        double Pace { get; set; }
         string Place { get; set; }
         DateTime Start { get; set; }
     }
